// examples/openapi-ts-axios/src/client/types.gen.ts
export type Category = {
  id?: number;
  name?: string;
};

export type User = {
  email?: string;
  firstName?: string;
  id?: number;
  lastName?: string;
  password?: string;
  phone?: string;
  /**
   * User Status
   */
  userStatus?: number;
  username?: string;
};

export type Tag = {
  id?: number;
  name?: string;
};

export type Pet = {
  category?: Category;
  id?: number;
  name: string;
  photoUrls: Array<string>;
  /**
   * pet status in the store
   */
  status?: 'available' | 'pending' | 'sold';
  tags?: Array<Tag>;
};